pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onebit-llm"
version = "0.1.0"
description = "1-bit / ternary LLM training and inference (Python rewrite)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "huggingface-hub>=0.20",
    "pyarrow>=14",
    "pandas>=2.0",
    "tqdm",
    "python-dotenv",
    "safetensors",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["onebit_llm*"]
